chore: removed some std:: prefixes from source file for readability of long function arguments chore: misc formatting

+++ source/includes/thmc/two_phase_flow/HeterogeneityAndRateAwareSaturationFunctions.h
@@ -115,7 +115,7 @@ class HeterogeneityAndRateAwareSaturationFunctions {
         double Bcp( Element<dim>* const e ) const;
         double Bcp_low( Element<dim>* const e ) const;
         VectorVariable<dim> InitializeVelocity( Element<dim>* const e ) const;
-        double krw_parallel( Element<dim>* const e, size_t direction ) const;
+        double krw_parallel( Element<dim>* const e, int direction ) const;
         double krw_crossflow( Element<dim>* const e ) const;
         double krw_parallel_at( Element<dim>* const e, double Sw, size_t direction ) const;
         double krw_crossflow_at( Element<dim>* const e, double Sw  ) const;
